Add initial travis-ci integration

+++ ci/prepare.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# Save trace setting
+XTRACE=$(set +o | grep xtrace)
+set -o xtrace
+
 # Read variables
 source ovn-scale.conf
 
@@ -8,11 +12,13 @@ sudo apt-get update -y
 sudo apt-get install -y apt-transport-https ca-certificates python-dev libffi-dev libssl-dev gcc make binutils
 sudo apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
 
-if [ ! -f /etc/apt/sources.list.d/docker.list ] ; then
-    sudo su -c 'echo "deb https://apt.dockerproject.org/repo ubuntu-trusty main" > /etc/apt/sources.list.d/docker.list' 
-    sudo apt-get update -y
-    sudo apt-get purge -y lxc-docker 
-    sudo apt-get install -y apparmor
+if [ "$INSTALLDOCKER" == "True" ] ; then
+    if [ ! -f /etc/apt/sources.list.d/docker.list ] ; then
+        sudo su -c 'echo "deb https://apt.dockerproject.org/repo ubuntu-trusty main" > /etc/apt/sources.list.d/docker.list' 
+        sudo apt-get update -y
+        sudo apt-get purge -y lxc-docker
+        sudo apt-get install -y apparmor
+    fi
 fi
 
 # Setup ssh
@@ -97,21 +103,26 @@ fi
 EOF
 sudo /etc/init.d/ssh restart
 
-# Install the docker engine
-sudo apt-get install -y docker-engine
-sudo service docker start
+if [ "$INSTALLDOCKER" == "True" ] ; then
+    # Install the docker engine
+    sudo apt-get install -y docker-engine
+    sudo service docker start
 
-# Create a docker group and add $OVNUSER user to this group
-EXISTING_DOCKER=$(cat /etc/group | grep docker)
-if [ "$EXISTING_DOCKER" == "" ]; then
-    sudo groupadd docker
-fi
-EXISTING_DOCKER_USER=$(cat /etc/group | grep docker | grep $OVNUSER)
-if [ "$EXISTING_DOCKER_USER" == "" ]; then
-    sudo usermod -aG docker $OVNUSER
-    if [ "$OVNSUDO" == "" ] ; then
-        echo "WARNING: The docker group was created and the $OVNUSER user added to this group."
-        echo "         Please reboot the box, log back in, and re-run $0."
-        exit 1
+    # Create a docker group and add $OVNUSER user to this group
+    EXISTING_DOCKER=$(cat /etc/group | grep docker)
+    if [ "$EXISTING_DOCKER" == "" ]; then
+        sudo groupadd docker
+    fi
+    EXISTING_DOCKER_USER=$(cat /etc/group | grep docker | grep $OVNUSER)
+    if [ "$EXISTING_DOCKER_USER" == "" ]; then
+        sudo usermod -aG docker $OVNUSER
+        if [ "$OVNSUDO" == "" ] ; then
+            echo "WARNING: The docker group was created and the $OVNUSER user added to this group."
+            echo "         Please reboot the box, log back in, and re-run $0."
+            exit 1
+        fi
     fi
 fi
+
+# Restore xtrace
+$XTRACE
